Removed the old log_name_dialog.py file. Moved the LogName class to log.py.
